requirements.txt
pillow>=10.2.0
PyMuPDF>=1.23.17